refactor: move Information Card
