+++ components/lookup/people/InformationCard.tsx
